+++ src/platform_impl/linux/wayland/keyboard.rs
@@ -404,12 +404,12 @@ fn keysym_to_vkey(keysym: u32) -> Option<VirtualKeyCode> {
 
 impl ModifiersState {
     pub(crate) fn from_wayland(mods: keyboard::ModifiersState) -> ModifiersState {
-        ModifiersState {
-            shift: mods.shift,
-            ctrl: mods.ctrl,
-            alt: mods.alt,
-            logo: mods.logo,
-        }
+        let mut m = ModifiersState::empty();
+        m.set(ModifiersState::SHIFT, mods.shift);
+        m.set(ModifiersState::CTRL, mods.ctrl);
+        m.set(ModifiersState::ALT, mods.alt);
+        m.set(ModifiersState::LOGO, mods.logo);
+        m
     }
 }